% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{create_session}
\alias{create_session}
\title{Create a treinus_session object}
\usage{
create_session(base_url, cookies, response, team_id = NULL)
}
\description{
Create a treinus_session object
}
\keyword{internal}
