% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{extract_teams_from_page}
\alias{extract_teams_from_page}
\title{Extract teams from team selection page}
\usage{
extract_teams_from_page(response)
}
\description{
Extract teams from team selection page
}
\keyword{internal}
