% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{print.treinus_pending_login}
\alias{print.treinus_pending_login}
\title{Print method for pending login}
\usage{
\method{print}{treinus_pending_login}(x, ...)
}
\arguments{
\item{x}{A treinus_pending_login object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for pending login
}
