% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{resolve_athlete_id}
\alias{resolve_athlete_id}
\title{Resolve athlete_id from argument or environment}
\usage{
resolve_athlete_id(athlete_id)
}
\description{
Resolve athlete_id from argument or environment
}
\keyword{internal}
