% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{resolve_team_id}
\alias{resolve_team_id}
\title{Resolve team_id from argument, session, or environment}
\usage{
resolve_team_id(team_id, session)
}
\description{
Resolve team_id from argument, session, or environment
}
\keyword{internal}
