% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{resolve_team_selection}
\alias{resolve_team_selection}
\title{Resolve team selection to team ID}
\usage{
resolve_team_selection(teams, team)
}
\description{
Resolve team selection to team ID
}
\keyword{internal}
