% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_cache_info}
\alias{treinus_cache_info}
\title{Get cache information}
\usage{
treinus_cache_info()
}
\value{
A tibble with columns: team_id, athlete_id, exercise_id, size_kb, modified
}
\description{
Returns information about cached exercise analysis data.
}
\examples{
\dontrun{
treinus_cache_info()
}

}
\seealso{
\code{\link[=treinus_cache_dir]{treinus_cache_dir()}}, \code{\link[=treinus_clear_cache]{treinus_clear_cache()}}
}
