% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_cache_path}
\alias{treinus_cache_path}
\title{Get cache file path for an exercise}
\usage{
treinus_cache_path(team_id, athlete_id, exercise_id)
}
\arguments{
\item{team_id}{Team ID}

\item{athlete_id}{Athlete ID}

\item{exercise_id}{Exercise ID}
}
\value{
Character string with cache file path
}
\description{
Get cache file path for an exercise
}
\keyword{internal}
