% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{treinus_config}
\alias{treinus_config}
\title{Get Treinus configuration}
\usage{
treinus_config()
}
\value{
A list with configuration details
}
\description{
Returns current configuration including credential status.
}
\examples{
treinus_config()

}
