% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_extract_records}
\alias{treinus_extract_records}
\title{Extract records from exercise analysis}
\usage{
treinus_extract_records(analysis)
}
\arguments{
\item{analysis}{Analysis data from \code{\link[=treinus_get_exercise_analysis]{treinus_get_exercise_analysis()}}}
}
\value{
A tibble with time-series records (GPS, heart rate, etc.)
}
\description{
Helper function to extract and tidy the time-series records from
exercise analysis data.
}
\examples{
\dontrun{
session <- treinus_auth()
analysis <- treinus_get_exercise_analysis(session, 57, 50, 2994)
records <- treinus_extract_records(analysis)
}

}
