% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{treinus_has_credentials}
\alias{treinus_has_credentials}
\title{Check if Treinus credentials are configured}
\usage{
treinus_has_credentials()
}
\value{
Logical indicating if credentials are set
}
\description{
Check if Treinus credentials are configured
}
\examples{
treinus_has_credentials()

}
