% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{treinus_session_valid}
\alias{treinus_session_valid}
\title{Check if a Treinus session is still valid}
\usage{
treinus_session_valid(session)
}
\arguments{
\item{session}{A treinus_session object from \code{\link[=treinus_auth]{treinus_auth()}}}
}
\value{
Logical indicating if the session is still valid
}
\description{
Check if a Treinus session is still valid
}
