% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_get_exercises}
\alias{treinus_get_exercises}
\title{Get last exercises for an athlete}
\usage{
treinus_get_exercises(session, athlete_id)
}
\arguments{
\item{session}{A treinus_session object from \code{\link[=treinus_auth]{treinus_auth()}}}

\item{athlete_id}{Integer ID of the athlete}
}
\value{
A tibble with exercise data including columns like \code{id_exercise_done},
\code{genre_name}, \code{start}, \code{distance}, \code{total_elapsed_time}, etc.
}
\description{
Retrieves the most recent exercises/workouts done by an athlete.
}
\examples{
\dontrun{
session <- treinus_auth()
exercises <- treinus_get_exercises(session, athlete_id = 50)
}

}
