% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{treinus_get_exercises_batch}
\alias{treinus_get_exercises_batch}
\title{Get exercises for multiple athletes}
\usage{
treinus_get_exercises_batch(session, athlete_ids, .progress = TRUE)
}
\arguments{
\item{session}{A treinus_session object from \code{\link[=treinus_auth]{treinus_auth()}}}

\item{athlete_ids}{Integer vector of athlete IDs}

\item{.progress}{Show progress bar? Default TRUE.}
}
\value{
A tibble with exercise data for all athletes
}
\description{
Retrieves the most recent exercises for a vector of athlete IDs.
}
\examples{
\dontrun{
session <- treinus_auth()
exercises <- treinus_get_exercises_batch(session, athlete_ids = 1:50)
}

}
